requirements.txt
python-telegram-bot[job-queue]>=20.0
sqlalchemy>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
google-cloud-storage>=2.13.0
packaging>=23.0